Fix editable menu item issue and redo the storage of of order data in thfireestore
