+++ src/lib/components/menu/editable/EditableMenuItem.svelte
@@ -1,10 +1,10 @@
 <script>
-import { createEventDispatcher } from 'svelte';
+import { createEventDispatcher, onMount, tick } from 'svelte';
 
 import EditableMenuItemOptions from './EditableMenuItemOptions.svelte';
 import tippy from '$lib/tippy';
 import Button from '$lib/components/input/Button.svelte';
-import Input from '../input/Input.svelte';
+import Input from '$lib/components/input/Input.svelte';
 import EditableMenuItemFilters from './EditableMenuItemFilters.svelte';
 
 export let item;
@@ -38,11 +38,12 @@ $: changed =
 
 $: removeDeletedOptions(allOptions);
 
-function removeDeletedOptions(allOptions) {
-    for (let option of item.options) {
-        if (!allOptions.includes(option)) {
-            save();
-        }
+async function removeDeletedOptions(allOptions) {
+    const allOptionIds = allOptions.map((option) => option.id);
+    // Wait for other state updates, specifically those in the autocomplete chooser which will update "options"
+    await tick();
+    if (item.options.some((option) => !allOptionIds.includes(option.id))) {
+        save();
     }
 }
 
@@ -70,47 +71,50 @@ function beforeUnload(e) {
 
 <div class="item" class:out-of-stock={!stock} bind:this={itemElement}>
     <img src="/menuItems/{image}" alt="Picture of {image}" />
-    <label>
-        Picture:
-        <Input bind:value={image} --font-size="16px" />
-    </label>
-    <label>
-        Name:
-        <Input bind:value={name} --font-size="16px" />
-    </label>
-    <label>
-        Price:
-        <Input bind:value={price} type="number" step="0.01" min="0" />
-    </label>
-    <EditableMenuItemOptions
-        popperBoundingElement={itemElement}
-        bind:selectedOptions={options}
-        options={allOptions} />
-    <EditableMenuItemFilters
-        popperBoundingElement={itemElement}
-        bind:selectedFilters={filters}
-        filters={allFilters} />
-
-    <label class="in-stock-label">
-        In Stock:
-        <input type="checkbox" bind:checked={stock} />
-    </label>
-
-    <div class="button-row">
-        <span use:tippy={!changed ? 'Make a change to save' : undefined}>
-            <Button on:click={save} disabled={!changed}>Save</Button>
+    <section class="inputs">
+        <span>
+            <Input label="Picture:" bind:value={image} --font-size="16px" />
         </span>
-        <span
-            use:tippy={item.stock
-                ? 'Item must be out of stock before deletion'
-                : undefined}>
-            <Button
-                disabled={item.stock}
-                on:click={() => dispatch('delete', item.id)}>
-                Delete
-            </Button>
+        <span>
+            <Input label="Name:" bind:value={name} --font-size="16px" />
         </span>
-    </div>
+        <span>
+            <Input label="Price:" bind:value={price} type="number" step="0.01" min="0" />
+        </span>
+    </section>
+    <section>
+        <EditableMenuItemOptions
+            popperBoundingElement={itemElement}
+            bind:selectedOptions={options}
+            options={allOptions} />
+        <EditableMenuItemFilters
+            popperBoundingElement={itemElement}
+            bind:selectedFilters={filters}
+            filters={allFilters} />
+    </section>
+    <section>
+        <span class="in-stock-label">
+            In Stock:
+            <input type="checkbox" bind:checked={stock} />
+        </span>
+    </section>
+    <section>
+        <div class="button-row">
+            <span use:tippy={!changed ? 'Make a change to save' : undefined}>
+                <Button on:click={save} disabled={!changed}>Save</Button>
+            </span>
+            <span
+                use:tippy={item.stock
+                    ? 'Item must be out of stock before deletion'
+                    : undefined}>
+                <Button
+                    disabled={item.stock}
+                    on:click={() => dispatch('delete', item.id)}>
+                    Delete
+                </Button>
+            </span>
+        </div>
+    </section>
 </div>
 
 <style>
@@ -132,27 +136,30 @@ function beforeUnload(e) {
 .item > img {
     width: 100%;
 }
-label {
+section {
+    margin: 5px 0px;
+    width: 100%;
+}
+section.inputs > span {
     padding: 10px;
     /* background-color: rgba(253, 253, 253, 0.5); */
     display: flex;
-    width: 100%;
     flex-flow: row nowrap;
     align-items: center;
     gap: 5px;
     color: rgba(0, 0, 0, 0.75);
-}
-label:hover {
-    cursor: pointer;
+    align-self: center;
 }
 .in-stock-label {
     align-self: center;
     width: auto;
+    padding: 5px;
 }
 .button-row {
     display: flex;
     flex-flow: row nowrap;
     gap: 10px;
-    align-self: center;
+    align-items: center;
+    justify-content: center;
 }
 </style>
