+++ src/lib/components/menu/editable/EditableMenuItemFilters.svelte
@@ -2,32 +2,53 @@
 import { slide } from 'svelte/transition';
 import { fade } from 'svelte/transition';
 import { flip } from 'svelte/animate';
-import AutocompleteChooser from '../input/AutocompleteChooser.svelte';
-import Option from './Option.svelte';
+import AutocompleteChooser from '$lib/components/input/AutocompleteChooser.svelte';
+import Option from '$lib/components/menu/Option.svelte';
 
-export let selectedOptions;
-export let options = [];
+export let selectedFilters;
+export let filters = {};
 export let popperBoundingElement;
 
 let chooser;
+let selectedOptions = filtersToOptions(filters).filter(({id}) => filters[id]);
+
+$: filterObjects = filtersToOptions(filters);
+const getFilters = () => filters;
+$: {
+    // unselectedFilters is a duplicate of filters with all filters set to false
+    const unselectedFilters = {};
+    Object.keys(getFilters()).forEach((k) => unselectedFilters[k] = false);
+    selectedFilters = {...unselectedFilters, ...optionsToFilters(selectedOptions)};
+}
+
+function filtersToOptions(filters) {
+    return Object.keys(filters).map((filterName) => { return {id: filterName, name: filterName}; });
+}
+
+function optionsToFilters(options) {
+    const filters = {};
+    for (let option of options) {
+        filters[option.name] = true;
+    }
+    return filters;
+}
 
 function remove(optionIndex) {
     chooser.remove(optionIndex);
 }
-const outOfStockMessage =
-    'This option is marked out of stock. It won\'t appear on the menu.';
-const removeOptionMessage = 'Remove this option';
+
+const removeOptionMessage = 'Remove this filter';
 </script>
 
 <div class="item-options" transition:slide|local>
     <AutocompleteChooser
         {popperBoundingElement}
-        placeholder="Search options"
-        options={options}
+        placeholder="Search filters"
+        options={filterObjects}
         bind:this={chooser}
         bind:selectedOptions
         let:option>
-        <Option {option} iconName={option.stock ? 'check' : 'close'}/>
+        <Option {option}/>
     </AutocompleteChooser>
     {#if selectedOptions.length > 0}
         <div class="options-list">
@@ -35,14 +56,13 @@ const removeOptionMessage = 'Remove this option';
                 <div animate:flip={{ duration: 250 }}>
                     <Option
                         {option}
-                        iconName={option.stock ? 'check' : 'close'}
-                        hoverMessage={option.stock ? removeOptionMessage : outOfStockMessage}
+                        hoverMessage={removeOptionMessage}
                         on:click={() => remove(option.id)}/>
                 </div>
             {/each}
         </div>
     {:else}
-        <p in:fade|local={{ duration: 250, delay: 250 }}>No options selected</p>
+        <p in:fade|local={{ duration: 250, delay: 250 }}>No filters selected</p>
     {/if}
 </div>
